+++ R/to_parsed_case_internal.R
@@ -1,13 +1,15 @@
-#' Internal parser, which is relevant for preprocessing, parsing and parsingoptions
+#' Internal parser, which is relevant for preprocessing, parsing and parsing options
 #'
 #' @param string A string.
-#' @param parsingoption An integer that will determine the parsingoption.
+#' @param parsing_option An integer that will determine the parsing option.
 #' \itemize{
 #'  \item{1: \code{RRRStudio -> RRR_Studio}}
 #'  \item{2: \code{RRRStudio -> RRRS_tudio}}
 #'  \item{3: parses at the beginning like option 1 and the rest like option 2.}
 #'  \item{4: parses at the beginning like option 2 and the rest like option 1.}
-#'  \item{any other integer: no parsing"}
+#'  \item{5: parses like option 1 but suppresses "_" around non alpha-numeric characters.
+#'  In this way this option suppresses splits and resulting case conversion after these characters.}
+#'  \item{any other integer <= 0: no parsing"}
 #'  }
 #'  
 #' @return A character vector separated by underscores, containing the parsed string.
@@ -17,7 +19,11 @@
 #'
 #' @importFrom magrittr "%>%"
 #'
-to_parsed_case_internal <- function(string, parsingoption = 1L){
+to_parsed_case_internal <- function(string, parsing_option = 1L){
+  ### input checking
+  if(parsing_option >= 6L){
+    stop("parsing_option must be 1,2,3,4,5 or <= 0 for no parsing.")
+  }
   ### preprocessing:
   # catch everything that should be handled like underscores
   # (only spaces by default)
@@ -64,26 +70,27 @@ to_parsed_case_internal <- function(string, parsingoption = 1L){
   
   ### applying parsing functions  
   # case: 1 RRRStudioSStudioStudio -> RRR_Studio_S_Studio_Studio
-  if(parsingoption == 1){
+  if(parsing_option == 1 | parsing_option == 5){
     string <- parsing_functions[["parse1_pat_cap_smalls"]](string)
     string <- parsing_functions[["parse2_pat_caps2"]](string)
     string <- parsing_functions[["parse3_pat_cap_lonely"]](string)
-    string <- parsing_functions[["parse4_separate_non_characters"]](string)}
+    string <- parsing_functions[["parse4_separate_non_characters"]](string)
+  }
   # case: 2 RRRStudioSStudioStudio -> RRRS_tudio_SS_tudio_Studio
-  if(parsingoption == 2){
+  if(parsing_option == 2){
     string <- parsing_functions[["parse2_pat_caps2"]](string)
     string <- parsing_functions[["parse1_pat_cap_smalls"]](string)
     string <- parsing_functions[["parse3_pat_cap_lonely"]](string)
     string <- parsing_functions[["parse4_separate_non_characters"]](string)}
   # case:3 RRRStudioSStudioStudio -> RRR_Studio_SS_tudio_Studio
-  if(parsingoption == 3){
+  if(parsing_option == 3){
     string <- parsing_functions[["parse5_pat_cap_smalls_first"]](string)
     string <- parsing_functions[["parse2_pat_caps2"]](string)
     string <- parsing_functions[["parse1_pat_cap_smalls"]](string)
     string <- parsing_functions[["parse3_pat_cap_lonely"]](string)
     string <- parsing_functions[["parse4_separate_non_characters"]](string)}
   # case:4 RRRStudioSStudioStudio -> RRRS_tudio_S_Studio_Studio
-  if(parsingoption == 4){
+  if(parsing_option == 4){
     string <- parsing_functions[["parse6_pat_caps2_first"]](string)
     string <- parsing_functions[["parse1_pat_cap_smalls"]](string)
     string <- parsing_functions[["parse2_pat_caps2"]](string)
@@ -94,6 +101,9 @@ to_parsed_case_internal <- function(string, parsingoption = 1L){
   string <- string %>%
     purrr::map_chr(~ stringr::str_replace_all(.x, "_+", "_")) %>% 
     purrr::map_chr(~ stringr::str_replace_all(.x, "^_|_$", ""))
+  if(parsing_option == 5){
+    string <- stringr::str_replace_all(string, "_(?![:alnum:])|(?<![:alnum:])_", "")
+  }
   # return
   string
 }